test(topic): separate topic in module and add tests
